requirements.txt
pure-eval==0.2.2
debugpy==1.6.7.post1
tokenizers==0.13.3
requests==2.31.0
cachetools==5.3.1
attrs==23.1.0
gitdb==4.0.10
toolz==0.12.0
huggingface-hub==0.16.4
importlib-metadata==6.6.0
prompt-toolkit==3.0.39
fonttools==4.42.1
watchdog==3.0.0
stack-data==0.6.2
filelock==3.12.3
urllib3==2.0.3
mdurl==0.1.2
sympy==1.12
pymongo==4.4.0
idna==3.4
smmap==5.0.0
click==8.1.3
gensim==4.3.1
tzlocal==4.3
zipp==3.15.0
deep-translator==1.11.4
dnspython==2.3.0
parso==0.8.3
PyJWT==2.7.0
smart-open==6.3.0
platformdirs==3.10.0
diffusers==0.20.2
Pympler==1.0.1
pickleshare==0.7.5
tornado==6.3.2
typing_extensions==4.6.3
ipykernel==6.25.1
streamlit==1.29.0
tenacity==8.2.2
kiwisolver==1.4.5
networkx==3.1
wcwidth==0.2.6
comm==0.1.4
markdown-it-py==3.0.0
validators==0.20.0
mpmath==1.3.0
Pygments==2.15.1
numpy==1.24.3
jedi==0.19.0
transformers==4.33.1
protobuf==4.23.3
altair==5.0.1
fsspec==2023.9.0
nest-asyncio==1.5.7
toml==0.10.2
charset-normalizer==3.1.0
bcrypt==4.0.1
scikit-learn==1.2.2
ipython==8.14.0
Pillow==9.5.0
matplotlib-inline==0.1.6
six==1.16.0
tensorflow
contourpy==1.1.0
accelerate==0.22.0
traitlets==5.9.0
Jinja2==3.1.2
threadpoolctl==3.1.0
firebase-admin==6.3.0
google-generativeai
jupyter_core==5.3.1
joblib==1.2.0
streamlit-antd-components==0.3.2
pydeck==0.8.1b0
streamlit==1.31.0
asttokens==2.2.1
executing==1.2.0
torch==2.0.1
seaborn==0.12.2
regex==2023.6.3
jsonschema==4.17.3
tqdm==4.65.0
matplotlib==3.7.2
blinker==1.6.2
extra-streamlit-components==0.1.56
PyYAML==6.0
colorama==0.4.6
openai
GitPython==3.1.31
nltk==3.8.1
fpdf==1.7.2
scipy==1.10.1
cycler==0.11.0
tzdata==2023.3
pyarrow==12.0.1
certifi==2023.5.7
rich==13.4.2
safetensors==0.3.3
psutil==5.9.5
decorator==5.1.1
pyzmq==25.1.1
gTTS==2.5.0
pyparsing==3.0.9
backcall==0.2.0
pandas==2.0.2
MarkupSafe==2.1.3
packaging==23.1
jupyter_client==8.3.1
